+++ cmd/aws-sso/ecs_client_cmd.go
@@ -32,9 +32,7 @@ import (
 	"github.com/synfinatic/gotable"
 )
 
-type EcsListCmd struct {
-	Auth string `kong:"help='Full HTTP Authorization token to use for ECS Server',env='AWS_CONTAINER_AUTHORIZATION_TOKEN'"`
-}
+type EcsListCmd struct{}
 
 type EcsLoadCmd struct {
 	// AWS Params
@@ -44,18 +42,15 @@ type EcsLoadCmd struct {
 	Profile   string `kong:"short='p',help='Name of AWS Profile to assume',predictor='profile',xor='account,role'"`
 
 	// Other params
-	Auth    string `kong:"help='Full HTTP Authorization token to use for ECS Server',env='AWS_CONTAINER_AUTHORIZATION_TOKEN'"`
-	Port    int    `kong:"help='TCP port of aws-sso ECS Server',env='AWS_SSO_ECS_PORT',default=4144"` // SEE ECS_PORT in ecs_cmd.go
-	Slotted bool   `kong:"short='s',help='Load credentials in a unique slot using the ProfileName as the key'"`
+	Port    int  `kong:"help='TCP port of aws-sso ECS Server',env='AWS_SSO_ECS_PORT',default=4144"` // SEE ECS_PORT in ecs_cmd.go
+	Slotted bool `kong:"short='s',help='Load credentials in a unique slot using the ProfileName as the key'"`
 }
 
 type EcsProfileCmd struct {
-	Auth string `kong:"help='Full HTTP Authorization token to use for ECS Server',env='AWS_CONTAINER_AUTHORIZATION_TOKEN'"`
-	Port int    `kong:"help='TCP port of aws-sso ECS Server',env='AWS_SSO_ECS_PORT',default=4144"`
+	Port int `kong:"help='TCP port of aws-sso ECS Server',env='AWS_SSO_ECS_PORT',default=4144"`
 }
 
 type EcsUnloadCmd struct {
-	Auth    string `kong:"help='Full HTTP Authorization token to use for ECS Server',env='AWS_CONTAINER_AUTHORIZATION_TOKEN'"`
 	Port    int    `kong:"help='TCP port of aws-sso ECS Server',env='AWS_SSO_ECS_PORT',default=4144"`
 	Profile string `kong:"short='p',help='Name of AWS Profile to unload',predictor='profile'"`
 }
@@ -74,7 +69,15 @@ func (cc *EcsLoadCmd) Run(ctx *RunContext) error {
 }
 
 func (cc *EcsProfileCmd) Run(ctx *RunContext) error {
-	c := client.NewECSClient(ctx.Cli.Ecs.Profile.Port, ctx.Cli.Ecs.Profile.Auth)
+	clientCert, err := ctx.Store.GetEcsSslCert()
+	if err != nil {
+		return err
+	}
+	bearerToken, err := ctx.Store.GetEcsBearerToken()
+	if err != nil {
+		return err
+	}
+	c := client.NewECSClient(ctx.Cli.Ecs.Profile.Port, bearerToken, clientCert)
 
 	profile, err := c.GetProfile()
 	if err != nil {
@@ -82,7 +85,7 @@ func (cc *EcsProfileCmd) Run(ctx *RunContext) error {
 	}
 
 	if profile.ProfileName == "" {
-		return fmt.Errorf("No profile loaded in ECS Server.")
+		return fmt.Errorf("no profile loaded in ECS Server")
 	}
 
 	profiles := []ecs.ListProfilesResponse{
@@ -92,7 +95,15 @@ func (cc *EcsProfileCmd) Run(ctx *RunContext) error {
 }
 
 func (cc *EcsUnloadCmd) Run(ctx *RunContext) error {
-	c := client.NewECSClient(ctx.Cli.Ecs.Unload.Port, ctx.Cli.Ecs.Unload.Auth)
+	clientCert, err := ctx.Store.GetEcsSslCert()
+	if err != nil {
+		return err
+	}
+	bearerToken, err := ctx.Store.GetEcsBearerToken()
+	if err != nil {
+		return err
+	}
+	c := client.NewECSClient(ctx.Cli.Ecs.Unload.Port, bearerToken, clientCert)
 
 	return c.Delete(ctx.Cli.Ecs.Unload.Profile)
 }
@@ -120,14 +131,30 @@ func ecsLoadCmd(ctx *RunContext, awssso *sso.AWSSSO, accountId int64, role strin
 	}
 
 	// do something
-	c := client.NewECSClient(ctx.Cli.Ecs.Load.Port, ctx.Cli.Ecs.Load.Auth)
+	clientCert, err := ctx.Store.GetEcsSslCert()
+	if err != nil {
+		return err
+	}
+	bearerToken, err := ctx.Store.GetEcsBearerToken()
+	if err != nil {
+		return err
+	}
+	c := client.NewECSClient(ctx.Cli.Ecs.Load.Port, bearerToken, clientCert)
 
 	log.Debugf("%s", spew.Sdump(rFlat))
 	return c.SubmitCreds(creds, rFlat.Profile, ctx.Cli.Ecs.Load.Slotted)
 }
 
 func (cc *EcsListCmd) Run(ctx *RunContext) error {
-	c := client.NewECSClient(ctx.Cli.Ecs.Profile.Port, ctx.Cli.Ecs.List.Auth)
+	clientCert, err := ctx.Store.GetEcsSslCert()
+	if err != nil {
+		return err
+	}
+	bearerToken, err := ctx.Store.GetEcsBearerToken()
+	if err != nil {
+		return err
+	}
+	c := client.NewECSClient(ctx.Cli.Ecs.Profile.Port, bearerToken, clientCert)
 
 	profiles, err := c.ListProfiles()
 	if err != nil {